- Remove unused sarc/core/models/cluster - Move sarc.core.models.validators into sarc.validators - Remove empty sarc.core.models - Move run_command from sarc.core.utils into sarc.utils - Remove now empty sarc.core.utils - Move sarc.core.scraping into sarc.scraping
