+++ sarc/scraping/prometheus.py
@@ -6,11 +6,11 @@ from sqlmodel import Session, col, select
 
 from sarc.cache import Cache, CacheEntry
 from sarc.config import ClusterConfig, config
-from sarc.core.scraping import series
 from sarc.db.cluster import SlurmClusterDB
 from sarc.db.job import JobStatisticDB, SlurmJobDB
 from sarc.db.runstate import get_parsed_date, set_parsed_date
 from sarc.models.job import SlurmState
+from sarc.scraping import series
 from sarc.traces import trace_decorator
 
 logger = logging.getLogger(__name__)
