+++ scripts/prometheus/restore.py
@@ -6,11 +6,11 @@ from dataclasses import dataclass
 from datetime import datetime
 
 from pydantic import BaseModel
+from sarc.client.job import JobStatistics, Statistics, _jobs_collection
 from simple_parsing import ArgumentParser, field
 from tqdm import tqdm
 
-from sarc.client.job import JobStatistics, Statistics, _jobs_collection
-from sarc.core.models.validators import datetime_utc
+from sarc.validators import datetime_utc
 
 logger = logging.getLogger("prometheus_dump.restore")
