+++ tests/functional/jobs/test_func_job_statistics.py
@@ -1,9 +1,9 @@
 import pytest
 from sqlmodel import select
 
-from sarc.core.scraping.series import JOB_STATISTICS_METRIC_NAMES, compute_job_statistics
 # from sarc.client.job import SlurmJob, get_job
 from sarc.db.job import SlurmJobDB
+from sarc.scraping.series import JOB_STATISTICS_METRIC_NAMES, compute_job_statistics
 
 BASE_ELAPSED_TIME = 60 * 60 * 12
