+++ tests/unittests/jobs/test_series.py
@@ -5,7 +5,7 @@ import pandas
 import pytest
 from pandas import DataFrame
 
-from sarc.core.scraping.series import compute_job_statistics_from_dataframe
+from sarc.scraping.series import compute_job_statistics_from_dataframe
 
 
 def _generate_df(rows, delta=30):
